requirements.txt
RPi.GPIO
serial
